master取り込み前

+++ Controller.cpp
@@ -48,24 +48,24 @@ void Controller::update(){
                     pre_conData.ButtonState = conData.ButtonState;
 
                     conData.ButtonState = 0, conData.LJoyX = 0, conData.LJoyY = 0, conData.RJoyX = 0, conData.RJoyY = 0;
-                    conData.ButtonState |= receive_data[0] - 0x20;
-                    conData.ButtonState |= (receive_data[1] - 0x20) << 6;
-                    conData.ButtonState |= (receive_data[2] - 0x20) << 12;
+                    conData.ButtonState |= (unsigned int)(receive_data[0] - 0x20);
+                    conData.ButtonState |= (unsigned int)(receive_data[1] - 0x20) << 6;
+                    conData.ButtonState |= (unsigned int)(receive_data[2] - 0x20) << 12;
                 
-                    conData.LJoyX |= (receive_data[3] - 0x20);
-                    conData.LJoyX |= ((receive_data[4] - 0x20) & 0x03) << 6;
+                    conData.LJoyX |= (unsigned int)(receive_data[3] - 0x20);
+                    conData.LJoyX |= (unsigned int)((receive_data[4] - 0x20) & 0x03) << 6;
                     conData.LJoyX = abs(conData.LJoyX - 0xFF);
 
-                    conData.LJoyY |= ((receive_data[4] - 0x20) & 0x3C) >> 2;
-                    conData.LJoyY |= ((receive_data[5] - 0x20) & 0x0F) << 4;
+                    conData.LJoyY |= (unsigned int)((receive_data[4] - 0x20) & 0x3C) >> 2;
+                    conData.LJoyY |= (unsigned int)((receive_data[5] - 0x20) & 0x0F) << 4;
                     conData.LJoyY = abs(conData.LJoyY - 0xFF);
 
-                    conData.RJoyX |= ((receive_data[5] - 0x20) & 0x30) >> 4;
-                    conData.RJoyX |= ((receive_data[6] - 0x20) & 0x3F) << 2;
+                    conData.RJoyX |= (unsigned int)((receive_data[5] - 0x20) & 0x30) >> 4;
+                    conData.RJoyX |= (unsigned int)((receive_data[6] - 0x20) & 0x3F) << 2;
                     conData.RJoyX = abs(conData.RJoyX - 0xFF);
 
-                    conData.RJoyY |= (receive_data[7] - 0x20);
-                    conData.RJoyY |= ((receive_data[8] - 0x20) & 0x03) << 6;
+                    conData.RJoyY |= (unsigned int)(receive_data[7] - 0x20);
+                    conData.RJoyY |= (unsigned int)((receive_data[8] - 0x20) & 0x03) << 6;
                     conData.RJoyY = abs(conData.RJoyY - 0xFF);
                 }
             }
